Taking an aquery output and converting it to a compilation database.
aquery does not provide information about how to process header files,
we fudge that here by assuming it would be the same as the correpsonding
*.cc file and emit the same compile action for the header.
Have to add -x c++ as the *.h name is ambiguous.

This does not work for header-only. And unfortunately, we're
entirely out of luck with generated *.cc and headers:

  * third_party/proto/kythe/analysis.pb.{h,cc}
  * third_party/proto/kythe/storage.pb.{h,cc}
  * verilog/CST/verilog_nonterminals_foreach-gen.inc
  * common/lsp/lsp-protocol.h
  * common/util/generated_verible_build_version.h
  * verilog/parser/verilog_parse_interface.h
  * verilog/parser/verilog_token_enum.h
  * verilog/tools/kythe/verilog_extractor_indexing_fact_type_foreach-gen.inc